notification-service/pyproject.toml
[tool.poetry]
name = "notification-service"
version = "0.1.0"
description = ""
authors = ["Higor Santos <higor.dantas@ccc.ufcg.edu.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
pydantic = "^2.6.1"
python-dotenv = "^1.0.1"
aiokafka = "^0.10.0"
chevron = "^0.14.0"
jsonschema = "^4.21.1"

[tool.poetry.scripts]
start = "notification_service.main:run"
doc = "asyncapi generate fromTemplate ./notification_service/api_specification/specification.yaml @asyncapi/html-template -o ./docs"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
